Getting the image sizes when generating the images

+++ cli/src/image/mod.rs
@@ -5,6 +5,13 @@ use std::path::PathBuf;
 pub struct Image {
     pub extension: String,
     pub dir: PathBuf,
-    pub raw_height: u16,
-    pub raw_width: u16,
+    pub raw_height: u32,
+    pub raw_width: u32,
+    // pub sizes: Vec<ImageSize>,
 }
+
+// #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
+// pub struct ImageSize {
+//     width: u32,
+//     height: u32,
+// }
